Rename LaunchpadLedger to LedgerConnector
